Push almost works, the pull has still some issues.
The Push works except that double occuring material are not recognized yet. So when pushing the same material twice you end up with two differnt materials with an equal set of parameters and name.
The Pull has still some issues. Most parameters for concrete are beeing pulled but some are not. The Concrete strengthe for both cubes and cylinders are 0.